+++ library/src/main/java/org/dbtools/android/commons/ui/navigation/ViewModelNav.kt
@@ -20,13 +20,14 @@ import kotlinx.coroutines.flow.asStateFlow
 interface ViewModelNav {
     val navigationActionFlow: StateFlow<NavigationAction?>
 
-    fun navigate(route: NavRoute, popBackStack: Boolean = false)
-    fun navigate(routes: List<NavRoute>)
-    fun navigate(route: NavRoute, navOptions: NavOptions)
-    fun navigate(route: NavRoute, optionsBuilder: NavOptionsBuilder.() -> Unit)
+    fun navigate(typeSafeRoute: Any, popBackStack: Boolean = false)
+    fun navigate(typeSafeRoutes: List<Any>)
+    fun navigate(typeSafeRoute: Any, navOptions: NavOptions)
+    fun navigate(typeSafeRoute: Any, optionsBuilder: NavOptionsBuilder.() -> Unit)
+    fun popBackStack(typeSafeRoute: Any? = null, inclusive: Boolean = false)
+    fun popBackStackWithResult(resultValues: List<PopResultKeyValue>, typeSafeRoute: Any? = null, inclusive: Boolean = false)
+
     fun navigate(intent: Intent, options: Bundle? = null, popBackStack: Boolean = false)
-    fun popBackStack(popToRouteDefinition: NavRouteDefinition? = null, inclusive: Boolean = false)
-    fun popBackStackWithResult(resultValues: List<PopResultKeyValue>, popToRouteDefinition: NavRouteDefinition? = null, inclusive: Boolean = false)
 
     fun navigate(navigationAction: NavigationAction)
     fun resetNavigate(navigationAction: NavigationAction)
@@ -36,32 +37,32 @@ class ViewModelNavImpl : ViewModelNav {
     private val _navigatorFlow = MutableStateFlow<NavigationAction?>(null)
     override val navigationActionFlow: StateFlow<NavigationAction?> = _navigatorFlow.asStateFlow()
 
-    override fun navigate(route: NavRoute, popBackStack: Boolean) {
-        _navigatorFlow.compareAndSet(null, if (popBackStack) NavigationAction.PopAndNavigate(route) else NavigationAction.Navigate(route))
+    override fun navigate(typeSafeRoute: Any, popBackStack: Boolean) {
+        _navigatorFlow.compareAndSet(null, if (popBackStack) NavigationAction.PopAndNavigate(typeSafeRoute) else NavigationAction.Navigate(typeSafeRoute))
     }
 
-    override fun navigate(routes: List<NavRoute>) {
-        _navigatorFlow.compareAndSet(null, NavigationAction.NavigateMultiple(routes))
+    override fun navigate(typeSafeRoutes: List<Any>) {
+        _navigatorFlow.compareAndSet(null, NavigationAction.NavigateMultiple(typeSafeRoutes))
     }
 
-    override fun navigate(route: NavRoute, navOptions: NavOptions) {
-        _navigatorFlow.compareAndSet(null, NavigationAction.NavigateWithOptions(route, navOptions))
+    override fun navigate(typeSafeRoute: Any, navOptions: NavOptions) {
+        _navigatorFlow.compareAndSet(null, NavigationAction.NavigateWithOptions(typeSafeRoute, navOptions))
     }
 
-    override fun navigate(route: NavRoute, optionsBuilder: NavOptionsBuilder.() -> Unit) {
-        _navigatorFlow.compareAndSet(null, NavigationAction.NavigateWithOptions(route, navOptions(optionsBuilder)))
+    override fun navigate(typeSafeRoute: Any, optionsBuilder: NavOptionsBuilder.() -> Unit) {
+        _navigatorFlow.compareAndSet(null, NavigationAction.NavigateWithOptions(typeSafeRoute, navOptions(optionsBuilder)))
+    }
+
+    override fun popBackStack(typeSafeRoute: Any?, inclusive: Boolean) {
+        _navigatorFlow.compareAndSet(null, NavigationAction.Pop(typeSafeRoute, inclusive))
+    }
+
+    override fun popBackStackWithResult(resultValues: List<PopResultKeyValue>, typeSafeRoute: Any?, inclusive: Boolean) {
+        _navigatorFlow.compareAndSet(null, NavigationAction.PopWithResult(resultValues, typeSafeRoute, inclusive))
     }
 
     override fun navigate(intent: Intent, options: Bundle?, popBackStack: Boolean) {
         _navigatorFlow.compareAndSet(null, if (popBackStack) NavigationAction.PopAndNavigateIntent(intent, options) else NavigationAction.NavigateIntent(intent, options))
-    }
-
-    override fun popBackStack(popToRouteDefinition: NavRouteDefinition?, inclusive: Boolean) {
-        _navigatorFlow.compareAndSet(null, NavigationAction.Pop(popToRouteDefinition, inclusive))
-    }
-
-    override fun popBackStackWithResult(resultValues: List<PopResultKeyValue>, popToRouteDefinition: NavRouteDefinition?, inclusive: Boolean) {
-        _navigatorFlow.compareAndSet(null, NavigationAction.PopWithResult(resultValues, popToRouteDefinition, inclusive))
     }
 
     override fun navigate(navigationAction: NavigationAction) {